added seed.py + seeded data into db

+++ models/restaurant_pizza.py
@@ -2,11 +2,9 @@ from models.dbconfig import db
 
 class RestaurantPizza(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    price = db.Column(db.Float, nullable=False)  # Adjusted to use Float for price
+    price = db.Column(db.Integer, nullable=False)
     restaurant_id = db.Column(db.Integer, db.ForeignKey('restaurant.id'), nullable=False)
     pizza_id = db.Column(db.Integer, db.ForeignKey('pizza.id'), nullable=False)
-    
-    # Define the relationship with Restaurant and Pizza models
     restaurant = db.relationship('Restaurant', back_populates='pizzas')
     pizza = db.relationship('Pizza', back_populates='restaurants')
